+++ src/services/owncloudservice.cpp
@@ -345,12 +345,12 @@ void OwnCloudService::settingsConnectionTest( SettingsDialog *dialog )
                 }
                 else
                 {
-                    settingsDialog->setOKLabelData( 5, "note path and ownCloud path are equal", SettingsDialog::Warning );
+                    settingsDialog->setOKLabelData( 5, QString( "note path and ownCloud path are equal" ).arg( notesPath ), SettingsDialog::Warning );
                 }
             }
             else
             {
-                settingsDialog->setOKLabelData( 5, "note path not in ownCloud path", SettingsDialog::Failure );
+                settingsDialog->setOKLabelData( 5, QString( "note path\n(%1)\nnot in ownCloud path" ).arg( notesPath ).arg( localOwnCloudPath ), SettingsDialog::Failure );
             }
         }
         else