+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - only todo items (not calendar items) are now shown in your todo lists
 - you can now choose to hide completed todo items at the todo list dialog
 - the last selected todo list is now remembered at the todo list dialog
+- improved the completed check for todo items
 
 ## 0.60
 - if the `Ctrl` key is pressed in the note text edit to click a link the mouse cursor is now a pointing hand